+++ main.lua
@@ -8,6 +8,8 @@ Curse = mods["Klehrik-CurseHelper"].setup()
 
 PATH = _ENV["!plugins_mod_folder_path"].."/"
 
+banlist = mods["Klehrik-Lost_in_Orbit"].aphelion_banlist
+
 
 
 -- ========== Main ==========
@@ -21,7 +23,12 @@ Initialize(function()
     }
     for _, folder in ipairs(folders) do
         local names = path.get_files(PATH..folder)
-        for _, name in ipairs(names) do require(name) end
+        for _, name in ipairs(names) do
+            if  banlist[folder]
+            and (not Helper.table_has(banlist[folder], path.stem(name))) then
+                require(name)
+            end
+        end
     end
 
     -- Misc